requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
boto3==1.35.35
botocore==1.35.35
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-jazzmin==3.0.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
